Start on weather remote data source

+++ feature-weather/domain/src/main/kotlin/com/aidannemeth/weatherly/feature/weather/domain/model/Temperature.kt
@@ -3,4 +3,4 @@ package com.aidannemeth.weatherly.feature.weather.domain.model
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Temperature(val value: Double)
+data class Temperature(val value: Float)
